test(test-optimization): split vitest no-worker init coverage
