+++ integration-tests/vitest/vitest.core.spec.js
@@ -28,7 +28,6 @@ const {
   TEST_COMMAND,
   TEST_SOURCE_FILE,
   TEST_SOURCE_START,
-  TEST_SUITE_ID,
   TEST_IS_NEW,
   TEST_NAME,
   TEST_EARLY_FLAKE_ENABLED,
@@ -70,7 +69,6 @@ versions.forEach((version) => {
   describe(`vitest@${version}`, () => {
     let cwd, receiver, childProcess, testOutput
     const newerVitestIt = version === '1.6.0' ? it.skip : it
-    const latestVitestIt = version === 'latest' ? it : it.skip
     const noWorkerUnsupportedIt = process.env.DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT ? it.skip : it
 
     useSandbox([
@@ -324,612 +322,6 @@ versions.forEach((version) => {
       ])
 
       assert.strictEqual(code, 0)
-    })
-
-    newerVitestIt('strips Datadog NODE_OPTIONS from fork workers when no-worker init is enabled', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvent = events.find(event => event.type === 'test')
-
-          assert.ok(testEvent, `should have test event, got events: ${inspect(events.map(event => event.type))}`)
-          assert.strictEqual(testEvent.content.meta[TEST_STATUS], 'pass')
-          assert.strictEqual(testEvent.content.meta[TEST_IS_TEST_FRAMEWORK_WORKER], 'true')
-          assert.ok(testEvent.content.duration > 0, 'test duration should be positive')
-          assert.ok(testEvent.content.duration < 60_000_000_000, 'test duration should be under 60s')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS:
-              '--conditions=C:\\tools\\hook.js --no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'forks',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            EXPECT_DD_NODE_OPTIONS_STRIPPED: '1',
-            EXPECT_DD_NODE_OPTIONS_WINDOWS_PATH_PRESERVED: '1',
-            EXPECT_TEST_DURATION: '1',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-    })
-
-    latestVitestIt('strips Datadog NODE_OPTIONS from thread workers when no-worker init is enabled', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvents = events.filter(event => event.type === 'test')
-
-          assert.strictEqual(testEvents.length, 1, inspect(events.map(event => event.type)))
-          assert.strictEqual(testEvents[0].content.meta[TEST_NAME], 'vitest worker env sets DD_VITEST_WORKER')
-          assert.strictEqual(testEvents[0].content.meta[TEST_STATUS], 'pass')
-          assert.strictEqual(testEvents[0].content.meta[TEST_IS_TEST_FRAMEWORK_WORKER], 'true')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'threads',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            EXPECT_DD_NODE_OPTIONS_STRIPPED: '1',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-    })
-
-    newerVitestIt('strips Datadog NODE_OPTIONS from fork project workers when root pool is threads', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvent = events.find(event => event.type === 'test')
-
-          assert.ok(testEvent, `should have test event, got events: ${inspect(events.map(event => event.type))}`)
-          assert.strictEqual(testEvent.content.meta[TEST_STATUS], 'pass')
-          assert.strictEqual(testEvent.content.meta[TEST_IS_TEST_FRAMEWORK_WORKER], 'true')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'threads',
-            PROJECT_POOL_CONFIG: 'forks',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            EXPECT_DD_NODE_OPTIONS_STRIPPED: '1',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-    })
-
-    newerVitestIt('reports fork project tests when custom sequencer runs before no-worker setup', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvent = events.find(event => event.type === 'test')
-
-          assert.ok(testEvent, `should have test event, got events: ${inspect(events.map(event => event.type))}`)
-          assert.strictEqual(testEvent.content.meta[TEST_STATUS], 'pass')
-          assert.strictEqual(testEvent.content.meta[TEST_IS_TEST_FRAMEWORK_WORKER], 'true')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'threads',
-            PROJECT_POOL_CONFIG: 'forks',
-            CUSTOM_SEQUENCER: 'true',
-            CUSTOM_SEQUENCER_MARKER,
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            EXPECT_DD_NODE_OPTIONS_STRIPPED: '1',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-      assert.ok(testOutput.includes(CUSTOM_SEQUENCER_MARKER), `Got: ${inspect(testOutput)}`)
-    })
-
-    latestVitestIt('reports thread specs routed from fork projects without worker init', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testSessionEvents = events.filter(event => event.type === 'test_session_end')
-          const testEvents = events
-            .filter(event => event.type === 'test')
-            .map(event => event.content)
-            .filter(test => test.meta[TEST_SOURCE_FILE].startsWith(
-              'ci-visibility/vitest-tests/no-worker-suite-context-'
-            ))
-          const sourceFiles = testEvents.map(test => test.meta[TEST_SOURCE_FILE]).sort()
-
-          assert.strictEqual(testSessionEvents.length, 1, inspect(events.map(event => event.type)))
-          assert.deepStrictEqual(sourceFiles, [
-            'ci-visibility/vitest-tests/no-worker-suite-context-a-slow.mjs',
-            'ci-visibility/vitest-tests/no-worker-suite-context-b-fast.mjs',
-          ])
-          assert.strictEqual(new Set(testEvents.map(test => test.test_session_id.toString())).size, 1)
-          assert.strictEqual(new Set(testEvents.map(test => test.test_module_id.toString())).size, 1)
-          for (const testEvent of testEvents) {
-            assert.strictEqual(testEvent.meta[TEST_STATUS], 'pass')
-            assert.strictEqual(testEvent.meta[TEST_IS_TEST_FRAMEWORK_WORKER], 'true')
-            assert.strictEqual(
-              testEvent.test_session_id.toString(),
-              testSessionEvents[0].content.test_session_id.toString()
-            )
-          }
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/no-worker-suite-context-*.mjs',
-            POOL_CONFIG: 'forks',
-            PROJECT_POOL_CONFIG: 'forks',
-            PROJECT_THREAD_POOL_MATCH_GLOB: '**/no-worker-suite-context-b-fast.mjs',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-    })
-
-    newerVitestIt('strips Datadog NODE_OPTIONS from fork projects when root thread pool disables isolate', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvent = events.find(event => event.type === 'test')
-
-          assert.ok(testEvent, `should have test event, got events: ${inspect(events.map(event => event.type))}`)
-          assert.strictEqual(testEvent.content.meta[TEST_STATUS], 'pass')
-          assert.strictEqual(testEvent.content.meta[TEST_IS_TEST_FRAMEWORK_WORKER], 'true')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'threads',
-            POOL_NO_ISOLATE: '1',
-            PROJECT_POOL_CONFIG: 'forks',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            EXPECT_DD_NODE_OPTIONS_STRIPPED: '1',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-    })
-
-    newerVitestIt('parents no-worker test events to their own suite when fork modules overlap', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testSuiteEvents = events
-            .filter(event => event.type === 'test_suite_end')
-            .map(event => event.content)
-            .filter(testSuite => testSuite.meta[TEST_SOURCE_FILE].startsWith(
-              'ci-visibility/vitest-tests/no-worker-suite-context-'
-            ))
-          const testEvents = events
-            .filter(event => event.type === 'test')
-            .map(event => event.content)
-            .filter(test => test.meta[TEST_SOURCE_FILE].startsWith(
-              'ci-visibility/vitest-tests/no-worker-suite-context-'
-            ))
-          const testSuiteIdsBySourceFile = new Map(testSuiteEvents.map(testSuite => [
-            testSuite.meta[TEST_SOURCE_FILE],
-            testSuite[TEST_SUITE_ID].toString(),
-          ]))
-
-          assert.strictEqual(testSuiteEvents.length, 2, inspect(testSuiteEvents.map(testSuite => testSuite.resource)))
-          assert.strictEqual(testEvents.length, 2, inspect(testEvents.map(test => test.resource)))
-          for (const test of testEvents) {
-            assert.strictEqual(
-              test[TEST_SUITE_ID].toString(),
-              testSuiteIdsBySourceFile.get(test.meta[TEST_SOURCE_FILE])
-            )
-          }
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/no-worker-suite-context-*.mjs',
-            POOL_CONFIG: 'forks',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-    })
-
-    newerVitestIt('reports suite hook failures when no-worker init is enabled', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testSuiteEvent = events.find(event => event.type === 'test_suite_end')
-
-          assert.ok(
-            testSuiteEvent,
-            `should have test suite event, got events: ${inspect(events.map(event => event.type))}`
-          )
-          assert.strictEqual(testSuiteEvent.content.meta[TEST_STATUS], 'fail')
-          assert.strictEqual(testSuiteEvent.content.meta[ERROR_MESSAGE], 'failed before all')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/failed-suite-hook.mjs',
-            POOL_CONFIG: 'forks',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 1, testOutput)
-    })
-
-    newerVitestIt('preserves string setupFiles when no-worker init is enabled', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvent = events.find(event => event.type === 'test')
-
-          assert.ok(testEvent, `should have test event, got events: ${inspect(events.map(event => event.type))}`)
-          assert.strictEqual(testEvent.content.meta[TEST_STATUS], 'pass')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/uses-string-setup-file.mjs',
-            VITEST_SETUP_FILE: 'ci-visibility/vitest-tests/string-setup-file.mjs',
-            POOL_CONFIG: 'forks',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-    })
-
-    latestVitestIt('does not duplicate thread project events when no-worker init is enabled', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvents = events.filter(event => event.type === 'test')
-
-          assert.strictEqual(testEvents.length, 2, inspect(testEvents.map(event => event.content.meta)))
-          testEvents.forEach(testEvent => {
-            assert.strictEqual(testEvent.content.meta[TEST_NAME], 'vitest worker env sets DD_VITEST_WORKER')
-            assert.strictEqual(testEvent.content.meta[TEST_STATUS], 'pass')
-          })
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            SECOND_PROJECT_TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'forks',
-            PROJECT_POOL_CONFIG: 'forks',
-            PROJECT_UNNAMED: '1',
-            SECOND_PROJECT_POOL_CONFIG: 'threads',
-            SECOND_PROJECT_UNNAMED: '1',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-    })
-
-    newerVitestIt('ignores no-worker init when isolate is disabled', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testSessionEvent = events.find(event => event.type === 'test_session_end')
-          const testModuleEvent = events.find(event => event.type === 'test_module_end')
-
-          assert.ok(
-            testSessionEvent,
-            `should have test session event, got events: ${inspect(events.map(event => event.type))}`
-          )
-          assert.ok(
-            testModuleEvent,
-            `should have test module event, got events: ${inspect(events.map(event => event.type))}`
-          )
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'forks',
-            NO_ISOLATE: '1',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            EXPECT_DD_NODE_OPTIONS_PRESENT: '1',
-            DD_TRACE_DEBUG: 'true',
-            DD_TRACE_LOG_LEVEL: 'warn',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-      assert.match(
-        testOutput,
-        /DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT is ignored because Vitest isolate is disabled/
-      )
-    })
-
-    newerVitestIt('ignores no-worker init when a fork project disables isolate', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvent = events.find(event => event.type === 'test')
-
-          assert.ok(testEvent, `should have test event, got events: ${inspect(events.map(event => event.type))}`)
-          assert.strictEqual(testEvent.content.meta[TEST_STATUS], 'pass')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'threads',
-            PROJECT_POOL_CONFIG: 'forks',
-            PROJECT_NO_ISOLATE: '1',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            EXPECT_DD_NODE_OPTIONS_PRESENT: '1',
-            DD_TRACE_DEBUG: 'true',
-            DD_TRACE_LOG_LEVEL: 'warn',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-      assert.match(
-        testOutput,
-        /DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT is ignored because Vitest isolate is disabled/
-      )
-    })
-
-    newerVitestIt('ignores no-worker init when a fork project disables pool isolate', async () => {
-      const payloadsPromise = receiver.gatherPayloadsMaxTimeout(
-        ({ url }) => url === '/api/v2/citestcycle',
-        payloads => {
-          const events = payloads.flatMap(({ payload }) => payload.events)
-          const testEvent = events.find(event => event.type === 'test')
-
-          assert.ok(testEvent, `should have test event, got events: ${inspect(events.map(event => event.type))}`)
-          assert.strictEqual(testEvent.content.meta[TEST_STATUS], 'pass')
-        }
-      )
-
-      childProcess = exec(
-        './node_modules/.bin/vitest run',
-        {
-          cwd,
-          env: {
-            ...getCiVisAgentlessConfig(receiver.port),
-            NODE_OPTIONS: '--no-warnings --import dd-trace/register.js -r dd-trace/ci/init',
-            TEST_DIR: 'ci-visibility/vitest-tests/vitest-worker-env.mjs',
-            POOL_CONFIG: 'threads',
-            PROJECT_POOL_CONFIG: 'forks',
-            PROJECT_POOL_NO_ISOLATE: '1',
-            DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT: 'true',
-            EXPECT_DD_NODE_OPTIONS_PRESENT: '1',
-            DD_TRACE_DEBUG: 'true',
-            DD_TRACE_LOG_LEVEL: 'warn',
-            DD_SERVICE: undefined,
-          },
-        }
-      )
-
-      childProcess.stdout.on('data', data => { testOutput += data })
-      childProcess.stderr.on('data', data => { testOutput += data })
-
-      const [[exitCode]] = await Promise.all([
-        once(childProcess, 'exit'),
-        payloadsPromise,
-      ])
-
-      assert.strictEqual(exitCode, 0, testOutput)
-      assert.match(
-        testOutput,
-        /DD_EXPERIMENTAL_TEST_OPT_VITEST_NO_WORKER_INIT is ignored because Vitest isolate is disabled/
-      )
     })
 
     noWorkerUnsupportedIt('propagates test span context to HTTP requests and hooks during test execution', async () => {
